requirements.txt
numpy
scipy
matplotlib
tqdm
cupy-cuda12x